Make `Chart.Animation/animations/Tooltip` importable
