+++ .htmllintrc
@@ -1,5 +1,6 @@
 {
 	"indent-style": "tabs",
+	"line-end-style": false,
 	"attr-quote-style": "double",
 	"spec-char-escape": false,
 	"attr-bans": [
